Starting work to seperate roles from userstore
